+++ src/api/relationApi.ts
@@ -6,7 +6,10 @@ import { get } from './fetch';
 const baseUrl = `${apiUrl}/relation`;
 
 export const getTimeSeriesRelations = async (): Promise<Relation[]> => {
-    const getRelationsResponse = await get(baseUrl, { category: 'Time series' });
+    const getRelationsResponse = await get(baseUrl, {
+        category: 'Time series',
+        tags: `community:${COMMUNITY_SLUG}`,
+    });
 
     return deserializeResponse<Relation[]>(getRelationsResponse);
 };